fix bounds
